+++ examples/Overview/src/main.m
@@ -1,7 +1,7 @@
 /*
  
  main.m
- LAPickerView Overview Example
+ LAUPickerView Overview Example
  
  Copyright (cc) 2012 Luis Laugga.
  Some rights reserved, all wrongs deserved.