+++ CHANGELOG.md
@@ -19,7 +19,7 @@ Features:
 
 Examples:
 
-  - Overview example (LAPickerView linked to UIPickerView and vice-versa)
+  - Overview example (LAUPickerView linked to UIPickerView and vice-versa)
  
 Other: